get work manager modifications

+++ composeApp/src/commonMain/kotlin/domain/usecase/usecase/auth/LogoutUseCase.kt
@@ -1,25 +1,53 @@
 package domain.usecase.usecase.auth
 
+import data.network.request.ChangeAvailabilityRequest
 import domain.repository.IAuthRepository
+import domain.repository.IAvailabilityRepository
 import domain.usecase.BaseUseCase
 import io.github.aakira.napier.LogLevel
 import io.github.aakira.napier.Napier
 import io.ktor.http.HttpStatusCode
 import irancell.nwg.wfm.getSharedPref
+import utils.Availability
+import utils.AvailabilityObjectId
 
 
 class LogoutUseCase(
-    private val iAuthRepository: IAuthRepository
+    private val iAuthRepository: IAuthRepository,
+    private val iAvailabilityRepository: IAvailabilityRepository
 ) : BaseUseCase<Unit, Unit>() {
     override suspend fun run(params: Unit) {
-        Napier.log(LogLevel.ASSERT,"Logout", message = "")
+            Napier.log(LogLevel.ASSERT, "Logout", message = "")
+            val available = getSharedPref().getBool(Availability, false)
+            if (available) {
+                val pair = iAvailabilityRepository.changeAvailability(
+                    ChangeAvailabilityRequest(
+                        false,
+                        getSharedPref().getString(AvailabilityObjectId)?.toInt()
+                    )
+                )
 
-       val pair =  iAuthRepository.logout()
-        if(pair.first == HttpStatusCode.OK){
-            iAuthRepository.deleteAllTableDB()
-            getSharedPref().deleteAll()
+                if (pair.first == HttpStatusCode.OK) {
+                    val pairLogout = iAuthRepository.logout()
+                    if (pairLogout in HttpStatusCode.OK..HttpStatusCode.MultiStatus || pairLogout == HttpStatusCode.Forbidden) {
+                        getSharedPref().deleteAll()
+                        iAuthRepository.deleteAllTableDB()
+                        Napier.log(LogLevel.ASSERT, "Logout", message = "Availability")
+                    }
+                }
+            } else {
+                val pairLogout = iAuthRepository.logout()
+                if (pairLogout in HttpStatusCode.OK..HttpStatusCode.MultiStatus || pairLogout == HttpStatusCode.Forbidden) {
+                    getSharedPref().deleteAll()
+                    iAuthRepository.deleteAllTableDB()
+                    Napier.log(LogLevel.ASSERT, "Logout", message = "pairLogout")
 
-        }
+
+                }
+
+            }
+
+
 
     }
